move functional tests to more standard dir
